requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.1.6
django-admin==2.0.2
django-excel-base==1.0.5
django-excel-response2==3.0.3
django-fsm==2.8.1
django-six==1.0.5
djangorestframework==3.14.0
pytz==2022.7.1
screen==1.0.1
sqlparse==0.4.3
xlwt==1.3.0
